src/main/java/com/ym/projectManager/service/impl/CustomerServiceImpl.java
package com.ym.projectManager.service.impl;

import com.ym.projectManager.model.Customer;
import com.ym.projectManager.repository.CustomerRepository;
import com.ym.projectManager.service.CustomerService;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class CustomerServiceImpl implements CustomerService {

    private final CustomerRepository customerRepository;

    public CustomerServiceImpl(CustomerRepository customerRepository) {
        this.customerRepository = customerRepository;
    }

    @Override
    public List<Customer> getAllCustomers() {
        return customerRepository.findAll();
    }
}
